+++ src/helpers/languageOptions.js
@@ -1,0 +1,11 @@
+module.exports = languageOptions = [
+  { label: 'C#', name: 'C#', value: 'csharp' },
+  { label: 'CSS', name: 'CSS', value: 'css' },
+  { label: 'HTML', name: 'HTML', value: 'html' },
+  { label: 'Java', name: 'Java', value: 'java' },
+  { label: 'Javascript', name: 'Javascript', value: 'javascript' },
+  { label: 'JSON', name: 'JSON', value: 'json' },
+  { label: 'SQL', name: 'SQL', value: 'sql' },
+  { label: 'Python', name: 'Python', value: 'python' },
+  { label: 'XML', name: 'XML', value: 'xml' }
+];